pyproject.toml
[project]
name = "flashcard-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Kolja Sam",email = "github@koljapluemer.com"}
]
# readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "psycopg2-binary>=2.9.0",
    "django-simple-history>=3.4.0",
    "django-summernote>=0.8.20",
    "reportlab>=4.0.0",
    "python-decouple>=3.8",
    "dj-database-url>=2.1.0",
    "gunicorn>=21.0.0",
    "whitenoise>=6.6.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
